Removing warnings and making ball lose life when it reaches the whole paddle height

+++ Core/Ticker.cpp
@@ -10,11 +10,11 @@ Ticker::Ticker()
 
 void Ticker::Tick()
 {
-	DeltaTime = (SDL_GetTicks() - LastDelta) * (TARGET_FPS / 1000.0);
+	DeltaTime = (float)((SDL_GetTicks() - LastDelta) * (TARGET_FPS / 1000.0));
 	if (DeltaTime > TARGET_DELTA)
 	{
 		DeltaTime = TARGET_DELTA;
 	}
-	LastDelta = SDL_GetTicks();
+	LastDelta = (float)SDL_GetTicks();
 
 }
